Update admin resources and add trivia API integration
- Fixed Filament form schemas (Section component namespace)
- Added Trivia API service for question imports
- Added new migrations for site content and nullable user_id
- Added Shane's custom background SVG and logo
- Added puppy video assets (webm, mp4)
- Updated Vue components (animations, layouts)
- Enhanced auth and contest controllers
- Added comprehensive test coverage

+++ app/Http/Controllers/DashboardController.php
@@ -87,7 +87,7 @@ class DashboardController extends Controller
                 'total' => $totalSubmissions,
                 'correct' => $correctSubmissions,
                 'incorrect' => $totalSubmissions - $correctSubmissions,
-                'accuracy' => $totalSubmissions > 0 ? round(($correctSubmissions / $totalSubmissions) * 100, 1) : 0,
+                'accuracy' => $totalSubmissions > 0 ? (float) round(($correctSubmissions / $totalSubmissions) * 100, 1) : 0.0,
             ],
         ]);
     }
